Rename "get-tree.js" to "get-element-tree.js"
